Move functions unchanged to new file
Import from new file cache_tools.py
